split main.py into submodules for easier testing and readability.
