+++ src/regioinvent/workflows/__init__.py
@@ -1,0 +1,35 @@
+from regioinvent.workflows.lcia_methods import import_fully_regionalized_impact_method
+from regioinvent.workflows.regionalization import connect_ecoinvent_to_regioinvent
+from regioinvent.workflows.regionalization import create_consumption_markets
+from regioinvent.workflows.regionalization import first_order_regionalization
+from regioinvent.workflows.regionalization import format_trade_data
+from regioinvent.workflows.regionalization import regionalize_ecoinvent_with_trade
+from regioinvent.workflows.regionalization import second_order_regionalization
+from regioinvent.workflows.regionalization import spatialize_elem_flows
+from regioinvent.workflows.regionalization import write_regioinvent_to_database
+from regioinvent.workflows.spatialization import spatialize_my_ecoinvent
+from regioinvent.workflows.regionalization import change_aluminium_electricity
+from regioinvent.workflows.regionalization import change_cobalt_electricity
+from regioinvent.workflows.regionalization import change_electricity
+from regioinvent.workflows.regionalization import change_heat
+from regioinvent.workflows.regionalization import change_waste
+from regioinvent.workflows.regionalization import test_input_presence
+
+__all__ = [
+    "format_trade_data",
+    "write_regioinvent_to_database",
+    "connect_ecoinvent_to_regioinvent",
+    "change_electricity",
+    "change_aluminium_electricity",
+    "change_cobalt_electricity",
+    "change_waste",
+    "change_heat",
+    "test_input_presence",
+    "spatialize_my_ecoinvent",
+    "import_fully_regionalized_impact_method",
+    "regionalize_ecoinvent_with_trade",
+    "first_order_regionalization",
+    "create_consumption_markets",
+    "second_order_regionalization",
+    "spatialize_elem_flows",
+]
